Move content into lib dir.
